[split] Improve the Dtab API
Problem

The current API is asymmetric: apply() doesn't refer to the same thing
as update(). This is highly confusing and even dangerous. The
relationship between the base Dtab and the local Dtab is unclear and
confusing. It's a bad API. Further, Java usage of Dtab was littered
with MODULE$ references.

Solution

Deconflate 'base' from 'local'. Instead of Dtab() and Dtab.base, we
have instead Dtab.base and Dtab.local. These are accessed and updated
in the standard way. Dtab users (i.e. BindingFactory, a class in
ConfigBus) must now compose the Dtab explicitly:
Dtab.base++Dtab.local. Also get rid of Dtab.baseDiff() -- protocols
simply encode Dtab.local.

We also ensure that Dtabs are tested with a nonempty Dtab.base, so
that this does not accidentally leak.

RB_ID=363999