[split] twitter-server: improve WebHandler heuristic for returning html or not
Problem

`com.twitter.server.WebHandler` uses a very limited heuristic for
deciding whether or not to return the response as HTML or not.
This leads to non-browser clients receiving HTML responses.

Solution

Improve the heuristic to be based on whether or not the user-agent
string begins with "Mozilla". See [1] for details.

Result

Browsers should get HTML and non-browsers should get the non-HTML
versions.

[1] http://en.wikipedia.org/wiki/User_agent#Format_for_human-operated_web_browsers

RB_ID=449537