twitter-server: Improve client profile to take into account session health
Problem

The client profile only took into account the request path. That is, failures
to acquire a service or failing endpoints are not reflected in the success rate
and such failure modes are not surfaced in the /admin interface.

Solution

Include the number of unavailable endpoints as part of the profile. This should
adequately represent finagle's failure management (failure accrual, fail fast, etc).
Clients are still sorted w.r.t success rate.

RB_ID=529943