Fix up static site server_info
Provides a mkRoute method in AdminHttpServer that allows Route objects
(which internally use the package private HttpUtils.{Request,Response})
to be created with finagle.http.{Request,Response} instances as are
likely to be found in code that wants to override the protected 'routes'
method of AdminHttpServer.

Uses that to augment the server_info with the staticsite specific server_info data.

Replaces a pre-source birdcage review at https://reviewboard.twitter.biz/r/525887/

Jira: DECK-2479

RB_ID=541542