csl: Use snapshot versions in sbt while off master
Problem

It's a hassle to develop off of the `develop` branch in OSS, because
a given `develop` may depend on an unpublished `develop` in another
repository.

Solution

Although a long-term solution may entail getting [dbuild][0] working,
that's not yet satisfactory for internal twitter usage, since dbuild
assumes that there is one repository per [sbt project][1].

An interim solution is to check whether we're on the master branch,
and use a "-SNAPSHOT" suffix otherwise.

Result

If you want to develop on finagle, but depend on new changes in
util, checkout util/develop, `./sbt publish-local`, and then checkout
finagle/develop and just `./sbt compile`.  It should "just work".

[0]: http://typesafehub.github.io/dbuild
[1]: https://github.com/typesafehub/dbuild/issues/159

RB_ID=569590