twitter-server: add stack trace for AdminHttpHandler exceptions. fix ShadowAdminServerTest Problem
Stack traces are not included for exceptions thrown by AdminHttpHandlers which can make it difficult to track down for common exception types. In addition, the ShadowAdminServerTest was not actually running.

Solution

Use the logger API that prints the stack trace. In addition, to fix the ShadowAdminServerTest, call the main method which pointed out that the per host handler was not being loaded, so add that to the services resource.

Result

The stack trace is now included in the log. In addition, the ShadowAdminServerTest runs.

RB_ID=589446