twitter-server: Exposes a basic registry endpoint
Problem

We need to slurp the R* Registry data out of somewhere

Solution

Much like with metrics, we can expose a Registry endpoint.  When
the other change lands, we can switch over to using
GlobalRegistry.withRegistry, but explicitly passing a registry in
the constructor is fine for now.

Moves ThreadsHandler's isHtml method into HtmlUtils in preparation for
using it in a more general context.

Result

Folks can start using R* Registry

RB_ID=588580