twitter-server: Adds build.properties to Registry
Problem

We want to export build.properties as part of the registry

Solution

Adds build.properties to the registry when ServerInfoHandler is
created.  TwitterServer is a singleton, so we're guaranteed it's
only done once.

Also fixes a flaky test in LoggingHandler to get twitter-server
tests to pass.

Result

Registry exports build.properties data.

RB_ID=591968