[twitter-server] Stream admin/events page
Problem

The Sink exposes an iterator over Events, which we can render into an
HTML stream (we already stream for the JSON handler), but we don't do
this for the HTML admin page. Consequently, this requires the controller
to fully buffer the HTML which could be potentially large.

Solution

Incrementally render and write HTML as we iterate over Events in the
Sink.

RB_ID=608234