[flight-recorder] Download Sink in Trace Event format
Problem

Currently the event capture can be displayed as an HTML table or as raw
JSON. We want to better understand the types of events that will be
collected in order to design useful presentations of those events. In
the interim, however, we have no solution for users who want a more
sophisticated presentation.

Solution

Trace Viewer[1] is built into Chrome, and greatly improves upon the
table presentation. We provide an endpoint for downloading an event
capture in Trace Event[2] format, which can then be loaded into Trace
Viewer.

[1] https://github.com/google/trace-viewer
[2] https://docs.google.com/a/twitter.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/edit#heading=h.yr4qxyxotyw

RB_ID=618082