Fix problems with SBT and Travis CI builds, update Scala and SBT versions
Problem

SBT builds for Finagle and associated projects were failing for a number of
reasons, including third-party dependency versions that had been updated for
Pants but not SBT, new tests that are run differently by the SBT test runner,
and changes to JSON serialization that don't work the same on Scala 2.10 and
2.11. Our SBT and Scala versions were also out of date.

Also Travis CI's container-based infrastructure currently has a bug that causes
the build to crash when getting the local hostname with IPv4 on OpenJDK 7
(reported in #3120 on GitHub). Moving back to the standard Travis infrastructure
is a trivial change, so I've included it here.

Solution

I've updated the SBT and Scala versions throughout, and have updated third-party
dependency versions to match the configuration in 3rdparty (note that I've
followed the last update to scrooge/project/build.properties in not updating the
custom SBT script for Scrooge). TwitterServer now uses scala.xml, which is a
separate module in Scala 2.11, so I've added a conditional dependency in the SBT
build.

The new ExitGuard tests were failing on SBT because clients were being left open
in other test suites. I've daemonized all of these clients so that they don't
create exit guards.

In JsonSinkTest in TwitterServer I've desugared an implicit conversion
provided by a view bound because Scala 2.11 prohibits Object as the target of an
implicit conversion.

Because Jackson doesn't correctly serialize Scala case classes that are defined
locally in vals on Scala 2.11, I've moved three case classes out of vals and
into the enclosing objects, changing names where necessary to avoid collisions,
and making each case class private to the object.

I've also switched all projects with a custom SBT script back to the standard
Travis CI infrastructure.

Result

We're on the most recent versions of Scala and SBT, the SBT builds actually
work, and Travis CI won't fail spuriously on OpenJDK 7.

RB_ID=619988