Change how jvm/gc/msec is reported in the UI.
Problem:
It is confusing to display on the UI that you're GC'ing at a certain rate, when the number you're actually displaying is how much time you've spent in GC overall. It has led engineers to jump to wrong conclusions about their systems.

Solution:
Change from displaying as amount GC'd per ms to displaying it as total amount of time spent GC'ing.

Result:
Reduce the confusion.