[twitter-server] Properly pass close deadline to AdminHttpServer
Problem
--

TwitterServer does not pass the close deadline to the admin http server when
closing. As a result, on calls to /quitquitquit, the server often closes the
connection before sending a response. This would not be too bad if it did not cause
Aurora to consider the task unhealthy, and result in immediate termination via a
signal.

Solution
--

Properly pass the close deadline.

RB_ID=753483