twitter-server: Change AdminHttpServer Tests to Properly Execute
Problem

The tests which involve checking whether the AdminServer and the
ShadowAdminServer are correctly serving stats and then closing are
not being run properly. The created TwitterServer instances do not
have their main method being called.

Solution

Create the individual servers and then call the main method of each
instance to ensure that the checkServer method gets run for each test.
Also, the ShadowAdminServer does not listen for the /quitquitquit
message so send that message instead to the AdminServer

Result

The tests are being properly executed and now are passing.

RB_ID=770891