twitter-server: Rewrites the browser history on button toggle in histograms
Problem

We want users to be able to paste the histogram url to their buddies
after they make changes to the way the graph looked, but the url
didn't change.

Solution

Rewrite the browser history to push the url they would have had to
follow to get to the new world.

Result

The url in the browser window now reflects what you would expect it
to be if you had used that url to get to the page.  Also, the back
button works as expected, as if it were a one page app.

RB_ID=819474