util-stats: Unify StatsReceivers that delegate
Problem

StatsReceivers end up looking like trees, because they can have one
child (if they proxy another stats receiver) or many children (if
they are created with BroadcastStatsReceiver).  It's useful to know
the properties of the stats receivers that we delegate to.

Solution

Provide marker traits for every branch node of the tree that lets
it see its children, and a helper method to find the leaves of a
tree.

Result

It's easy to find the leaves of a StatsReceiver tree given any
StatsReceiver that's correctly marked.

RB_ID=819519