finagle/finatra/scrooge/twitter-server: bin/travisci updates and fixes
Problem

The order of building needs to be updated for twitter-server and scrooge
as building finagle should now happen after building the scrooge-sbt-plugin.
The finagle/finatra builds are complaining about not finding the 2.10 version of
the scrooge-generator.

Solution

Update the order of building in twitter-server and scrooge scripts.
Specifically, publish the scrooge-generator for 2.10 in the finagle
script.

RB_ID=824936
TBR=true