twitter-server: Remove AdminHttpServer#mkRoutex method
Problem

This method was created during the migration away from
direct usage of netty http types and is now redundant.

Solution

The AdminHttpServer#mkRoutex method is now functionally
equivalent to the AdminHttpServer#mkRoute method so the
former can be deleted and usages moved to the latter.

Result

Less tech debt!

RB_ID=835083