twitter-server: Make Server Summary Depend on Pending Stat
Problem

The Twitter Server Summary page contains a section which displays basic
statistics on running servers. It is using the presence of the "load"
gauge to determine whether to display a particular server within this
section. The "load" gauge was removed in Finagle 6.36.0 as it is a
duplicate of the "pending" gauge.

Solution

Switch the server section of the Summary page to use the "pending" gauge
instead of the "load" guage so that the graph displays again for the
appropriate set of servers.

RB_ID=898422