twitter-server: Endpoint to get Tunables for all ids
Problem

In addition to being able to get tunables for a given id,
we'd like an endpoint on ts-admin to see all Tunables for
all ids.

Solution

Handle GET requests to the /admin/tunables endpoint on
twitter-server, and add the page to the index under "process info".
Requests return a JSON string of all TunableMaps and their contents.

RB_ID=913237