twitter-server: Add slf4j-api logging implementation support for jdk14, log4j12, and logback.
Summary: Problem

When TwitterServer is updated to integrate the slf4j-api logging framework (in place
of util-logging) we will support the dynamic toggling of log levels in the admin
HTTP interface for 3 logging implementations: JUL, log4j12, and logback-classic.

Solution

This commit adds the base functionality for the mentioned 3 logging implementations
to support admin handlers which will allow for dynamic log level toggling -- however,
THIS COMMIT DOES NOT WIRE IN THESE CHANGES. There are additional (breaking API)
changes required in the base TwitterServer code base that will come in following commits.

Result

More groundwork for the integration of the slf4j-api into TwitterServer.

JIRA Issues: CSL-3923

Differential Revision: https://phabricator.twitter.biz/D97316