twitter-server/slf4j-jdk14: Apply log format in trait constructor
Summary: Problem

The application of the LogFormat to the JUL logging handlers
is done in the `premain` lifecycle of the server which is
after a fair amount of server constructor logging.

Solution

Move the application of the LogFormat to the constructor
of the trait which will apply the LogFormat to the configured
handlers earlier. This comes with the drawbacks that anyone
overrriding the log format in the trait will now not be able
to use flags for configuration.

In practice this is most likely not a general concern.

Differential Revision: https://phabricator.twitter.biz/D106534