twitter-server: Make StatsReceiver in Stats trait overridable
Summary: Problem

It is not currently possible to easily override the StatsReceiver
member defined in the c.t.server.Stats trait since it is defined
as a val and thus eagerly referenced during construction. Generally
members in traits should be specified as defs to allow extensions
to define the eagerness of instantiation. This makes it hard to
test server instances with a mock or dummy StatsReceiver (i.e,
an InMemoryStatsReceiver).

Solution

Change the `statsReceiver` val to be a def. The def by default
references a global singleton (`LoadedStatsReceiver`) so this
should not have any suprising behavior for users but does add
a method invocation.

Result

Easier to swap out the `c.t.server.Stats#statsReceiver` to be
able to assert collected stats on servers under test.

JIRA Issues: CSL-5578

TBR=true

Differential Revision: https://phabricator.twitter.biz/D113214