twitter-server: Add c.t.server.Hook#onExit lifecycle callback
Problem

There may be cases where a Hook needs to run in the `App#onExit`
lifecycle.

Solution

Add an `onExit` callback function which can be implemented by a
`c.t.server.Hook` and implement calling it in the Hooks trait
which loads and fires service-loaded hooks.

Result

A `c.t.server.Hook` can define logic which will be triggered during
the `onExit` lifecycle phase of an App.

Differential Revision: https://phabricator.twitter.biz/D198379