twitter-server: Make parsed build properties available through object
Problem

It would be great to allow for other code in TwitterServer to have
access to the parsed `build.properties` without needing to re-parse
the file as the `c.t.server.handler.ServerInfoHandler` already does this.

Solution

Move the logic to parse the `build.properties` into a separate object
with utility methods to expose access to the read properties. Update
the `ServerInfoHandler` to use this object utility.

Result

Other TwitterServer code can get access to the read `build.properties`
file without needing to re-parse the file.

JIRA Issues: CSL-6665

Differential Revision: https://phabricator.twitter.biz/D201207