Rename CHANGES to CHANGELOG.rst
Problem

Github renders reStructuredText (rst) files. While our changelog
is written in rst, because it does not have that as its file type
it is rendered as plain text.

Solution

Rename to CHANGELOG.rst. This uses changelog instead of changes,
which seems to be the more canonical name.

Result

Changelogs are rendered nicely in the Github UI.

Differential Revision: https://phabricator.twitter.biz/D206432