twitter-server: Do not warn when there is no build.properties to load
Problem/Solution

The `c.t.server.BuildProperties` is rather noisy when there is no `build.properties`
to load, i.e., when running integration tests that start a server.

This results in a `NullPointerException`, so we specifically change to handle that
case in loading the properties file to not output a log message. Thus the warning
should appear only when there is a problem loading an actual file.

Differential Revision: https://phabricator.twitter.biz/D229586