twitter-server: Add Details for Watch Admin Endpoint
Problem / Solution

The admin documentation contains little information on the `Watch` endpoint. It
can be used to view how a specific metric changes over time for a TwitterServer
instance, as well as to download data for a specific metric. Let's add some
additional detailed documentation for this endpoint.

JIRA Issues: CSL-7243

Differential Revision: https://phabricator.twitter.biz/D273549