twitter-server: Make /histograms{,.json} endpoints consistent
Problem

/histograms?h=$HISTO&fmt=raw returns json as a raw array, but /histograms.json
returns a big json object where the keys are the names of the histograms, and the
values are the raw array.  Having two different formats makes it harder to build
tools that can consume either of them.

Solution

Change the format to match the /histograms.json style.

Result

/admin/histograms can now output data that looks like:

{
  "jvm/gc/eden/pause_msec" : [
    {
      "lowerLimit" : 8,
      "upperLimit" : 9,
      "count" : 1
    },
    {
      "lowerLimit" : 9,
      "upperLimit" : 10,
      "count" : 1
    },
    {
      "lowerLimit" : 53,
      "upperLimit" : 54,
      "count" : 1
    }
  ]
}

JIRA Issues: CSL-7717

Differential Revision: https://phabricator.twitter.biz/D279779