util-app|twitter-server: Track duplicated flags and add a lint rule
Problem

It is possible to register the same Flag name in a `c.t.app.App` multiple times.
`c.t.app.App#flag` is an instance of `c.t.app.Flags` which stores Flags in a
`Map[String, Flag[_]]` keyed by Flag name where the last registered name wins.

The current logic prints a notice  to System.err when a duplicated Flag name is
registered, however, this can lead to unintended behavior for users with little
visibility into what is happening.

Solution

Track the registration of Flags with the same name and add a lint rule to
TwitterServer, `DuplicateFlagDefinitions` which will give visibility in the form
of a lint rule violation when multiple Flags with the same name are registered
in the app.

Result

Better visibility for users when multiple Flags with the same name are
registered in the app.

JIRA Issues: CSL-7829

Differential Revision: https://phabricator.twitter.biz/D314410