twitter-server: admin pages script order
Problem

We've empirically observed server instances that aren't able to render
the charts in /admin/metrics. The current working theory is that this is
a dependency issue in which google charts are loaded and the respective
callback is dispatched before `chart-renderer.js` is loaded.

Solution

Load `chart-renderer.js` before trying to load the google charts lib.
`ChartRenderer` depends on the the google charts libs, but it's only
used after the libs have successfully loaded.

Differential Revision: https://phabricator.twitter.biz/D345063