twitter-server: Restructure LoadBalancersHandler Logic
Problem / Solution

LoadBalancer data can be rendered as either JSON or HTML, and
all the logic for doing so is contained within the `LoadBalancerHandler`.
As we restructure endpoints within the Twitter Server Admin Interface, let's
separate out the action, from the data model, from the rendered content to
make it easier to work with moving forward.

JIRA Issues: CSL-9014

Differential Revision: https://phabricator.twitter.biz/D392168