[logging] lift slf4j-log4j12 deps from libraries to binaries
Problem
slfj is a logging facade with several implementations - which may be loaded at runtime. It is generally incorrect to include an implementation in a library. Instead, include just the API in library code and allow applications to depend on a specific implementation in the binary

Some libraries include logging implementations.

Solution

lift the implementation dependency from library to binary.

Result

a No-Op change

Differential Revision: https://phabricator.twitter.biz/D398671