twitter-server: flakey AdminThreadPoolFilter test
Problem:
AdminThreadPoolFilter checks `numCompletedTasks` on the admin thread pool at
the end of the test. `numCompletedTasks` is specified to be an approximation
in the JavaDoc.

Solution / Result:
Remove this specific assert. `AdminThreadPoolFilter.Pool` no longer has to be
packaged scoped either. The test already ensures that the filter shifts work
to another thread.

JIRA Issues: CSL-9087

Differential Revision: https://phabricator.twitter.biz/D400449