twitter-server: Allow failures on Scala 2.13.1 TravisCI builds
Problem

I was a bit overzealous removing the allow failures matrix in
the previous commit. We still need this as not all of finagle
is Scala 2.13.1 compatible. The issue is when building the
dependencies of twitter-server with Dodo, we first run an
`sbt ++SCALA_VERSION update` in the repo to be built. For
twitter-server this means when we try to build finagle for
Scala 2.13.1 we end up attempting to resolve Scala 2.12 artifacts.
The reason this is a problem is we cache in TravisCI by Scala
version and thus no Scala 2.12 artifacts will be retrievable
when we are building for Scala 2.13 and thus the TravisCI
build fails:

Solution

Until all of Finagle is Scala 2.13 compatible, we will allow the
2.13.1 builds to fail in TravisCI such that the build is not marked
as a complete failure or error.

Differential Revision: https://phabricator.twitter.biz/D445397