twitter-server: Better admin logging handler support
Problem

The logic to install a TwitterServer Admin logging handler for
dynamic configuration of log levels per logging implementation
suffers from a lack of specifity making it hard to detect when
the incorrect case of multiple handlers being present on the
classpath and thus loaded by `LoadService` arises.

Solution

Separate out loading of the logging handler to better detect
the case of multiple handlers. Add a lint rule which is in
violation when multiple are loaded. Make sure the TwitterServer
admin only attempts to install a handler when there is only
a single one specified.

JIRA Issues: CSL-9764

Differential Revision: https://phabricator.twitter.biz/D484965