Revert 'Allow AdminHttpServer to be started later'
This reverts commit d99e96d0a7a284d11dc3fdb99030c42c24dbe83c.

These changes were to help solve a lifecycle impedance mismatch between
`c.t.server.Stats#statsReceiver` and the Finatra framework. However,
it seems that solving that mismatch will be nigh impossible for legacy
reasons and because the hacks to do so make usage of the framework
very brittle and prone to error -- e.g., the path of the current
solution would mean users would need to know that accessing the
`c.t.server.Stats#statsReceiver` should only be done at a certain point
in the lifecycle. We will likely need to resort to documenting the
issue and providing clarity on how to deal with the inconsistencies
over a real fix. Since these AdminHttpServer changes were meant to
help solve the lifecycle issue, it is better to revert them since they
will not be used and may cause more harm than good if left.

Differential Revision: https://phabricator.twitter.biz/D500871