twitter-server: Add the current metric name as a url fragment in the metrics view
Problem

The metrics Watch endpoint will use the url fragment to go right to the
metrics you're interested in which is super convenient. However, you
essentially have to type this in by hand since it's never populated in
the browsers url bar.

Solution

Set the url fragment when we start to watch a metric. This makes copying
and sharing admin metrics pages super quick.

Differential Revision: https://phabricator.twitter.biz/D635343