twitter-server: Update histograms docs and admin page to use snake case
Problem

The histograms admin endpoint was updated in c90fa00caeba to use
snake_case for field names. The same change missed updating
documentation for the endpoint and the JavaScript that supports
histograms in the admin interface.

Solution

Updated docs and JS code to use snake_case field names for upper and
lower limits of histogram buckets.

Closes twitter/twitter-server#68

Signed-off-by: Lily He <lhe@twitter.com>

JIRA Issues: CSL-10945

Differential Revision: https://phabricator.twitter.biz/D667345