twitter-server: add name and namespace params to ExpressionHandler
Problem
Consumers of the /admin/metric/expressions.json endpoint would like to be able to grab a subset of the expression by name or namespace. There is no way to do so currently.

Solution
Add optional query params name and namespace which are used to filter the list of expressions returned by the /admin/metric/expressions.json endpoint.

JIRA Issues: CSL-10622

Differential Revision: https://phabricator.twitter.biz/D672594