twitter-server: lock-free metric store
# Problem

The implementation of `MetricStore` uses synchronized blocks to handle the refresh of metrics. I noticed that it becomes a significant contention point for Strato. JFR profiling session:

{F32332182}

# Solution

Optimize the implementation to use a lock-free mechanism.

# Notes

There's a minor behavior change that I believe isn't a problem: the previous implementation would block reads while the map was being refreshed but the new lock-free version allows other threads to read the stale map while the refresh is happening.

Differential Revision: https://phabricator.twitter.biz/D682705