finagle|finatra|scrooge|ts|util: Use sbt version in GH actions cache key
Problem

The builds were failing because pre-sbt upgrade, the existing sbt launch jar was
cached, which caused a conflict when trying to use the upgraded sbt launch jar
(which had the same name as the old one).

Solution

Add the sbt version to the cache key so that when the sbt version is upgraded,
the cache will be invalidated and everything will run smoothly.

JIRA Issues: CSL-11029

Differential Revision: https://phabricator.twitter.biz/D683985