twitter-server: responsive UI for admin stack module registry
Problem:
In /admin/clients and /admin/servers, when the text in the parameter table of
the Stack Module page gets too long, the entire table shifts to the bottom of
the page. This gives the false impression that the description of the module is
missing.

Solution:
Contain the panel in bootstrap grid columns. Break the long text so it wraps
around to the next line.

Result:
The table will always display on the right no matter the width of the screen.
The table width will grow and shrink depending on the screen width.

JIRA Issues: CSL-11037

Differential Revision: https://phabricator.twitter.biz/D686836