twitter-oss: Move templates to PULL_REQUEST_TEMPLATE folder
Problem:
GitHub needs the special keyword "closes" to automatically close PRs when a
commit is merged.

Solution:

Add "closes twitter/$REPOSITORY#PR_NUMBER" to the pull request template so PRs
from open source automatically close when the commit is merged in. This avoids
manually closing the PR when possible.

Also, move all the pull request templates to a PULL_REQUEST_TEMPLATE folder
because this is the recommended standard from GitHub

Differential Revision: https://phabricator.twitter.biz/D692645