twitter-server: Change the default behavior for rating of expressions
Problem

We use latching_style as a work-around so that callers don't need to also
understand whether each metric is a counter or not, and whether the counters
are latched or not.  However, in the long run we would like to switch to not
reporting rated-ness, so we can instead rely on the callers understanding the
metrics in a deeper way.  Currently, the "rated" style is the default, and it
will show the "true" latch style if you pass in "latching_style=true".

Solution

Update "not rated" to be the default.

Result

We will only rate metrics if it is appropriate for them to be rated (ie they
are unlatched) and latching_style=true is passed in.

JIRA Issues: CSL-11145

Differential Revision: https://phabricator.twitter.biz/D702897