{util,twitter-server}: Make StringExpression higher fidelity
Problem

StringExpression is currently incompatible with rating because we don't know
whether it's backed by a counter or not!

Solution

Add a boolean as a quick hack to let us know whether the metric is a counter
or not.  This API may change rapidly as we learn more about how it should be
used.

JIRA Issues: CSL-11715

Differential Revision: https://phabricator.twitter.biz/D832571