util-stats: more support for creating a Stat expression
**Problem/Solution**
Move `HistogramComponent` to its own class

**Problem/Solution**
Provide default percentiles in `HistogramComponent`

**Problem**
Percentile is not a component of `HistogramComponent`,  to create a
Stat expression, users need to pass either a `HistogramComponent`
or a percentile as `Double`.

**Solution**
Introduce `Percentile` as a new `HistogramComponent` and refactor
the API to create a Stat expression to take a `HistogramComponent`.
This makes it easier for us to introduce a new API to create a list of
Stat expressions with different `HistogramComponent` (will be done
in a separate phab).

JIRA Issues: CSL-11828

Differential Revision: https://phabricator.twitter.biz/D855103