Add a point to the AdminHttpServer to latch additional config hooks onto
twitter-server: Add marker Param to the Stack for AdminHttpServer

Problem/Solution:

The AdminHttpServer is not differentiated from any other service that it
administrates and is difficult to configure separately, should a user want to.

This change adds a Stack.Param that marks a Stack as being a Admin interface
should a user want to write additional configuration hooks that target this
Stack.Param.

JIRA Issues: CSL-11988

Differential Revision: https://phabricator.twitter.biz/D908894