util-stats, twitter-server: Add a ShortLived verbosity level
Problem

Some metrics are not valuable for a long time and it would be nice if
the backend metrics store could drop them early to avoid the cost of
longer term storage.

Solution

Add a ShortLived metrics verbosity level that can signal to the metrics
store through the metric metadata endpoint that this metrics is of limited
value for long duration.

Differential Revision: https://phabricator.twitter.biz/D954203