twitter-server: apply IndexView to all combined muxers
Problem

We only apply IndexView to routes added statically via
the `Admin` trait. This means that any routes which are
added via the global HttpMuxer can't take advantage of
the styles and templating inside of `IndexView`.

Solution

Move combine inside of `AdminHttpServer` and apply the filter
dynamically inside of it.

Result

`IndexView` is now applied to all the muxers combined in the
`AdminHttpServer`.

Differential Revision: https://phabricator.twitter.biz/D989672